Delete MessageSendingService and move its logic to interceptor
