+++ stream-chat-android-offline/src/main/java/io/getstream/chat/android/offline/experimental/plugin/factory/StreamOfflinePluginFactory.kt
@@ -32,7 +32,6 @@ import io.getstream.chat.android.offline.experimental.plugin.listener.ThreadQuer
 import io.getstream.chat.android.offline.experimental.plugin.listener.TypingEventListenerImpl
 import io.getstream.chat.android.offline.experimental.plugin.logic.LogicRegistry
 import io.getstream.chat.android.offline.experimental.plugin.state.StateRegistry
-import io.getstream.chat.android.offline.message.MessageSendingServiceFactory
 import io.getstream.chat.android.offline.repository.creation.builder.RepositoryFacadeBuilder
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.SupervisorJob
@@ -98,20 +97,16 @@ public class StreamOfflinePluginFactory(
         val stateRegistry = StateRegistry.getOrCreate(job, scope, userStateFlow, repos, repos.observeLatestUsers())
         val logic = LogicRegistry.getOrCreate(stateRegistry)
 
-        val messageSendingServiceFactory = MessageSendingServiceFactory.getOrCreate(
-            logic,
-            globalState,
-            scope,
-            repos,
-            config.uploadAttachmentsNetworkType
+        val sendMessageInterceptor = SendMessageInterceptorImpl(
+            context = appContext,
+            logic = logic,
+            globalState = globalState,
+            repos = repos,
+            scope = scope,
+            networkType = config.uploadAttachmentsNetworkType
         )
-
         val defaultInterceptor = DefaultInterceptor(
-            sendMessageInterceptor = SendMessageInterceptorImpl(
-                context = appContext,
-                logic = logic,
-                messageSendingService = messageSendingServiceFactory
-            )
+            sendMessageInterceptor = sendMessageInterceptor
         )
 
         val channelMarkReadHelper = ChannelMarkReadHelper(
@@ -127,10 +122,10 @@ public class StreamOfflinePluginFactory(
             addUserConnectedListener(chatDomainImpl::userConnected)
 
             addUserDisconnectedListener {
+                sendMessageInterceptor.cancelJobs() // Clear all jobs that are observing attachments.
                 stateRegistry.clear()
                 logic.clear()
                 globalState.clearState()
-                messageSendingServiceFactory.getAllServices().forEach { it.cancelJobs() }
             }
         }
